requirements.txt
Flask>=2.0
requests>=2.28
chromadb>=0.3.26
python-dotenv>=1.0.0
